+++ NEWS.md
@@ -39,7 +39,7 @@
 ### Changes
 
 - Adds compatibility with julia 0.6 development versions, including automated tests
-- Fixes Censored BUM π0 edge case when all p-values are equal to 1
+- Fixes Censored BUM π₀ edge case when all p-values are equal to 1
 - Updates test cases and reference values for `qValues`
 - Introduces `Compat` as a new dependency
 - Separates test coverage reports for different julia versions
@@ -92,7 +92,7 @@
 
 ### New Features
 
-- New π0 estimators: BUM, Censored BUM, Flat Grenander
+- New π₀ estimators: BUM, Censored BUM, Flat Grenander
 - New p-value adjustment methods: Forward Stop, Benjamini-Liu
 - BUM model for simulating p-value distributions
 - Fast isotonic regression
@@ -123,7 +123,7 @@
 
 ### New features
 
-- New π0 estimators: Two Step, Right Boundary
+- New π₀ estimators: Two Step, Right Boundary
 - New p-value adjustment methods: Benjamini-Hochberg Adaptive, Oracle, Sidak
 - Package precompilation